double -> float

+++ MathHelper.cpp
@@ -2,8 +2,8 @@
 
 #include "MathHelper.hpp"
 
-const double MathHelper::PC_TO_KM = 3.08567758129e13;
-const double MathHelper::SEC_PER_YEAR = 365.25 * 86400;
-const double MathHelper::DEG_TO_RAD = M_PI / 180.0;
-const double MathHelper::RAD_TO_DEG = 180.0 / M_PI;
-const double MathHelper::CONTANT_OF_GRAVITY = 6.672e-11;
+const float MathHelper::PC_TO_KM = 3.08567758129e13f;
+const float MathHelper::SEC_PER_YEAR = 365.25f * 86400;
+const float MathHelper::DEG_TO_RAD = (float)M_PI / 180.0f;
+const float MathHelper::RAD_TO_DEG = 180.0f / (float)M_PI;
+const float MathHelper::CONTANT_OF_GRAVITY = 6.672e-11f;
