+++ include/SDLWnd.hpp
@@ -55,7 +55,7 @@ protected:
 	double GetFOV() const;
 	SDL_Event m_event;
 
-	void DrawText(TTF_Font *pFont, TextCoords coords, int x, int y, const char* fmt, ...);
+	void DrawText(TTF_Font *pFont, TextCoords coords, float x, float y, const char* fmt, ...);
 
 	static Vec3D GetOGLPos(int x, int y);
 	static Vec2D GetWindowPos(GLfloat x, GLfloat y, GLfloat z);
